#19 - arreglos de ubicacion de archivos
