+++ app/__init__.py
@@ -27,9 +27,11 @@ def not_found(error):
 
 # Import a module / component using its blueprint handler variable (mod_auth)
 from app.authentication.controllers import mod_auth as auth_module
+from app.article.controllers import mod_art as art_module
 
 # Register blueprint(s)
 app.register_blueprint(auth_module)
+app.register_blueprint(art_module)
 # app.register_blueprint(xyz_module)
 # ..
