+++ app/authentication/controllers.py
@@ -12,9 +12,8 @@ from werkzeug import check_password_hash, generate_password_hash
 
 from app import db
 from app.authentication.constants import ReadRole, CommentRole, WriteRole
-from app.authentication.forms import LoginForm, SignupForm, ArticleCreateForm
-from app.authentication.forms import SectionCreateForm
-from app.authentication.models import User, Article, Section
+from app.authentication.forms import LoginForm, SignupForm
+from app.authentication.models import User
 
 mod_auth = Blueprint('auth', __name__, url_prefix='/auth')
 
@@ -35,38 +34,6 @@ def profile():
         return redirect(url_for('auth.signin'))
     else:
         return render_template('authentication/profile.html')
-
-
-@mod_auth.route('/create', methods=['GET', 'POST'])
-def article_create():
-    if 'email' not in session:
-        return redirect(url_for('sigin'))
-    user     = User.query.filter_by(email=session['email']).first()
-    username = user.username
-    article  = Article()
-    form     = ArticleCreateForm()
-    form.user_name.data = user.username
-    if form.validate_on_submit():
-        form.populate_obj(article)
-        db.session.add(article)
-        db.session.commit()
-        return redirect(url_for('index'))
-    return render_template('user/create.html', form=form, user=user, username=username)
-
-
-@mod_auth.route('/section/create', methods=['GET', 'POST'])
-def section():
-    form = SectionCreateForm()
-    section = Section()
-    person = Person.query.filter_by(email=session['email']).first()
-    name = person.firstname
-    if form.validate_on_submit():
-        form.populate_obj(section)
-        db.session.add(section)
-        db.session.commit()
-        return redirect(url_for('dashboard', name=name))
-    return render_template('cat_create.html', form=form)
-
 
 @mod_auth.route('/signup/', methods=['GET', 'POST'])
 def signup():
