+++ app/authentication/forms.py
@@ -3,23 +3,7 @@ from wtforms import TextField, PasswordField, SubmitField  # BooleanField
 from wtforms import TextAreaField, HiddenField
 from wtforms.ext.sqlalchemy.fields import QuerySelectField
 from wtforms.validators import Required, Email, EqualTo, Length, Regexp
-from models import db, User, Section
-
-strip_filter = lambda x: x.strip() if x else None
-
-def section_choice():
-    return Section.query.all()
-
-class ArticleCreateForm(Form):
-    title     = TextField('Title', [Required("Please enter a title")],
-                        filters=[strip_filter])
-    body      = TextAreaField('Body',[Required("Please enter a body")],
-                        filters=[strip_filter])
-    section   = QuerySelectField('Section', query_factory=section_choice )
-    user_name = HiddenField()
-
-class ArticleUpdateForm(Form):
-    id = HiddenField()
+from models import db, User
 
 
 class SignupForm(Form):
@@ -65,7 +49,3 @@ class LoginForm(Form):
         else:
             self.password.errors.append("Incorrect Password")
         return False
-
-class SectionCreateForm(Form):
-    name = TextField('Name', Length(min=1,max=240))
-    description = TextAreaField('Description')
